Bug Fixes, Punch Functionality and Login Enabled
Removed the EmployeeService and am going to go for a single user model instead. This way the EU manages all of their own data. Enabled authentication using firebase and implemented the register and reset password features.
Added the punch service and linked it to the FB database so punches can be added.
Started to work and retrieving the users punches from the data.